Add extra heating
This patch makes it possible to pass extra heating terms to grackle. It's similar to the volumetric/specific heating term added in more recent versions. Here the chenge is made in a minimal way, just enough to make the cxx_example work. It's not a thorough API update. We need it for AGORA GIZMO runs. This patch is slightly modified from the one kindly provided by @alessandro_lupi.